% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create map of fatal injuries}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer between 1 and 56 corresponding to the state number}

\item{year}{4-digit integer specifying the year}
}
\value{
None
}
\description{
This function displays a basic map of where fatal injuries suffered in motor
vehicle traffic crashes occured for a specific year in a specific state.
}
\note{
An invalid state number or your will result in an error. If no
accidents occured in the given state and your a message will be printed.
The data (csv.bz2) needs to be in the working directory.
}
