% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in FARS data for several years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of 4-digit integers specifying the years for
which the data should be loaded.}
}
\value{
Returns a list of data frames (tibbles) corresponding to the FARS
data sets for the input years
}
\description{
This function reads in multiple FARS data sets for several years.
}
\note{
The csv files need to be named correctly and need to be saved in the
working directory. If the i-th entry of the input years is invalid the i-th
value of the returned list will be NULL and a warning message will be given.
The data (csv.bz2) needs to be in the working directory.
}
