% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Display number of records per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of 4-digit integers specifying the years for
which to calculate the number of fatal injuries.}
}
\value{
This function returns a data frame (tibble) with the number of fatal
injuries per month (rows) for every year (columns) input.
}
\description{
This function summarizes the number of  fatal injuries suffered in motor
vehicle traffic crashes in each month for each year requested by the user.
}
\note{
Partial input checking is done by invoked functions. If no valid year
is given as input, an error results. The data (csv.bz2) needs to be in the working directory.
}
