% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A four-digit integer specifying the year}
}
\value{
Return the filename of the FARS data set for the input year.
}
\description{
This simple function creates the filename of a FARS data set given a
year specific
}
\note{
No input checking is done, so be sure to specify a valid year.
}
\examples{
make_filename(2014)

}
