% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create map of fatal injuries}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{year}{4-digit integer specifying the year}

\item{state_num}{Integer between 1 and 56 corresponding to the state number}
}
\value{
None
}
\description{
This function displays a basic map of where fatal injuries suffered in motor
vehicle traffic crashes occured for a specific year in a specific state.
}
\note{
An invalid state number or your will result in an error. If no
accidents occured in the given state and your a message will be printed.
}
\examples{
# Data needs to be in working directory for this example to work
fars_map_state(12, 2013)

}
