% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the (path and) filename of the
FARS data file to be loaded}
}
\value{
Returns the corresponding data frame (tibble) in R.
}
\description{
This functions reads in a data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System (FARS) saved as a
CSV file.
}
\note{
An error is invoced if no valid filename is given.
The data (csv.bz2) needs to be in the working directory.
}
